Rearranged controller code
